LayoutTests/platform/mac/accessibility/multiselect-list-reports-active-option-expected.txt


This tests that navigating in a multiselect list updates selection and the active selected option and sends a notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibleOne.isSelected is true
FAIL accessibleOne.isSelectedOptionActive should be true. Was false.
PASS accessibleTwo.isSelected is false
PASS accessibleTwo.isSelectedOptionActive is false
PASS accessibleThree.isSelected is false
PASS accessibleThree.isSelectedOptionActive is false
PASS accessibleOne.isSelected is false
PASS accessibleOne.isSelectedOptionActive is false
PASS accessibleTwo.isSelected is true
FAIL accessibleTwo.isSelectedOptionActive should be true. Was false.
PASS accessibleThree.isSelected is false
PASS accessibleThree.isSelectedOptionActive is false
PASS accessibleOne.isSelected is false
PASS accessibleOne.isSelectedOptionActive is false
PASS accessibleTwo.isSelected is true
PASS accessibleTwo.isSelectedOptionActive is false
PASS accessibleThree.isSelected is true
FAIL accessibleThree.isSelectedOptionActive should be true. Was false.
List notification: AXLayoutComplete
List notification: AXLayoutComplete
List notification: AXLayoutComplete
List notification: AXSelectedChildrenChanged
List notification: AXLayoutComplete
List notification: AXLayoutComplete
List notification: AXLayoutComplete
List notification: AXSelectedChildrenChanged
List notification: AXLayoutComplete
List notification: AXLayoutComplete
List notification: AXLayoutComplete

TEST COMPLETE
